add github action for target Linux, Mac and Windows
