+++ TRAPTACloud.desktop
@@ -1,0 +1,8 @@
+
+[Desktop Entry]
+Type=Application
+Name=TRAPTACloud
+Comment=TRAPTA Cloud
+Exec=TRAPTACloud
+Icon=traptacloud
+Categories=Office;
