+++ TRAPTACloud.pro
@@ -4,9 +4,17 @@ TEMPLATE = app
 QT += core network qml quick
 CONFIG += c++11
 CONFIG += release
-QMAKE_LFLAGS += -static 
 
-
+debug {
+    DESTDIR = bin/debug
+    MOC_DIR = build/lib/debug
+    OBJECTS_DIR = build/lib/debug
+}
+release {
+    DESTDIR = bin/release
+    MOC_DIR = build/lib/release
+    OBJECTS_DIR = build/lib/release
+}
 
 unix:!macx {
     QMAKE_LFLAGS += -no-pie
@@ -24,11 +32,11 @@ DEFINES += QT_DEPRECATED_WARNINGS
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-        main.cpp \
-    viewcontroller.cpp \
-    udplistener.cpp \
-    cloud.cpp \
-    trapta.cpp
+    src/main.cpp \
+    src/viewcontroller.cpp \
+    src/udplistener.cpp \
+    src/cloud.cpp \
+    src/trapta.cpp
 
 RESOURCES += resources.qrc \
 
@@ -36,7 +44,7 @@ RC_FILE = traptacloud.rc
 ICON = traptacloud.icns
 
 HEADERS += \
-    viewcontroller.h \
-    udplistener.h \
-    cloud.h \
-    trapta.h
+    src/viewcontroller.h \
+    src/udplistener.h \
+    src/cloud.h \
+    src/trapta.h
