+++ config.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Installer>
+    <Name>TRAPTA Cloud</Name>
+    <Version>1.0</Version>
+    <Title>TRAPTA Cloud</Title>
+    <Publisher>TRAPTA Project</Publisher>
+    <StartMenuDir>TRAPTA</StartMenuDir>
+    <TargetDir>@HomeDir@/trapte/traptacloud</TargetDir>
+</Installer>
