+++ package.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Package>
+    <DisplayName>TRAPTA Cloud</DisplayName>
+    <Description>TRAPTA Cloud</Description>
+    <Version>1.0</Version>
+    <ReleaseDate>2017-04-27</ReleaseDate>
+    <Default>true</Default>
+</Package>
